kafka-producer/requirements.txt
# Kafka Producer Dependencies
kafka-python==2.0.2
osmnx==2.0.7
geopandas==1.1.1
shapely==2.1.2